+++ test/parse_test.dart
@@ -1,5 +1,4 @@
 import 'package:test/test.dart';
-import 'package:xml/xml.dart';
 
 import 'assertions.dart';
 
@@ -205,10 +204,5 @@ void main() {
         assertFragmentParseError('<?processing', 'Expected name', 1);
       });
     });
-  });
-  test('parse deprecated', () {
-    // ignore: deprecated_member_use_from_same_package
-    final document = parse('<?xml version="1.0"?><schema/>');
-    expect(document, isA<XmlDocument>());
   });
 }