+++ CHANGELOG.md
@@ -1,12 +1,17 @@
 # Changelog
 
-## 5.4.0 (unpublished)
+## 6.0.0 (unpublished)
 
 * Dart 2.16 requirement.
 * Escape control characters (thanks to https://github.com/rspilker).
 * Add a predicate to pretty printer to insert a space character before self-closing elements (thanks to https://github.com/rspilker).
 * Add predicates to normalizer to trim leading and trailing whitespaces, as well as collapse consecutive whitespaces.
 * Expose `qualifiedName`, `localName`, `namespacePrefix` and `namespaceUri` for convenience on the named nodes.
+* Remove deprecated code:
+  * `parse(String input)`: use `XmlDocument.parse(String input)` or `XmlDocumentFragment.parse(String input)` instead.
+  * `XmlBuilder.build()`: use `XmlBuilder.buildDocument()` or `XmlBuilder.buildFragment()` instead.
+  * `XmlNormalizer.defaultInstance`: use `const XmlNormalizer()` instead.
+  
 
 ## 5.3.0